fix(web): unwrap server rejection reasons in hand-rolled error toasts
Five mutations built their toast from error.message, which for hono's
DetailedError is just the status line — a rejected status update read
as "400 Bad Request" instead of the server's reason. Promote
readServerErrorMessage out of transactions' checkout-issues into
lib/server-error so orders and the worker route can use it without
reaching into another feature.
